steramlining processing script
