+++ config/config.yaml
@@ -1,12 +1,26 @@
 
 samples: "config/samples.tsv"
-
+config: "config/config.yaml"
 
 # options are 'bwa' 'hisat2' and 'bowtie2'
 aligner: 'bwa'
 
+experiment_name: '2021_Sar11Pro_RNAseq_Project'
+
 htseq:
   mapping_qual_threshold: 10 # default is 10
+
+feature_types_to_keep: [
+'CDS',
+'rRNA',
+'tRNA',
+'sRNA'
+# 'RNase_P_RNA',
+# 'pseudogene',
+# 'region',
+# 'riboswitch',
+# 'tmRNA',
+]
 
 input: 
   raw_reads: /nfs/chisholmlab001/kve/2021_Sar11Pro_RNAseq_Project/data/input_data/raw_reads
